third_party/WebKit/LayoutTests/fast/dom/shadow/focus-shadowhost-display-none.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
div#host:focus { display: none; }
</style>
<div id="host"></div>
<script>
description('Check if shadow host with display:none CSS rule for :focus works.  crbug.com/482830');

var root = host.createShadowRoot();
var input = document.createElement('input');
root.appendChild(input);

host.tabIndex = 0;
host.tabStop = true;

function testFocusShadowHost() {
    debug('when shadow host itself is focused, it should match display:none, lose focus then becomes display:block again.');
    return new Promise(
        function(resolve) {
            host.focus();
            shouldBeEqualToString('window.getComputedStyle(host).display', 'none');
            shouldBe('document.activeElement', 'host');
            shouldBeNull('root.activeElement');

            function onBlur() {
                shouldBeEqualToString('window.getComputedStyle(host).display', 'block');
                shouldBe('document.activeElement', 'document.body');
                shouldBeNull('root.activeElement');
                host.removeEventListener('blur', onBlur);
                resolve();
            }
            host.addEventListener('blur', onBlur);
        });
}

function testFocusInsideShadowRoot() {
    debug('when shadow host with tabStop=false has focused element inside the shadow, it should also match display:none, then lose focus and become display:block again.');
    return new Promise(
        function(resolve) {
            host.tabStop = false;
            input.focus();
            shouldBeEqualToString('window.getComputedStyle(host).display', 'none');
            shouldBe('document.activeElement', 'host');
            shouldBe('root.activeElement', 'input');

            function onBlur() {
                shouldBeEqualToString('window.getComputedStyle(host).display', 'block');
                shouldBe('document.activeElement', 'document.body');
                shouldBeNull('root.activeElement');
                input.removeEventListener('blur', onBlur);
                resolve();
            }
            input.addEventListener('blur', onBlur);
        });
}

if (window.testRunner) {
    testFocusShadowHost().then(testFocusInsideShadowRoot).then(function(){ testRunner.notifyDone(); });
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}
</script>
